+++ examples/base/src/index.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { ApolloClient, createNetworkInterface } from 'apollo-client';
+import ApolloClient from 'apollo-boost';
 import { ApolloProvider } from 'react-apollo';
 
 import { App } from './App';
 
 const client = new ApolloClient({
-  networkInterface: createNetworkInterface({
-    uri: 'https://mpjk0plp9.lp.gql.zone/graphql',
-  }),
+  uri: 'https://mpjk0plp9.lp.gql.zone/graphql',
 });
 
 const WrappedApp = (