+++ src/Subscriptions.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as PropTypes from 'prop-types';
 import ApolloClient, { ApolloError } from 'apollo-client';
-import { Observable } from 'apollo-client/util/Observable';
+import { Observable } from 'apollo-link';
 
 import { DocumentNode } from 'graphql';
 import { ZenObservable } from 'zen-observable-ts';
@@ -17,9 +17,9 @@ export interface SubscriptionResult<TData = any> {
 }
 
 export interface SubscriptionProps<TData = any, TVariables = OperationVariables> {
-  query: DocumentNode;
+  subscription: DocumentNode;
   variables?: TVariables;
-  shouldResubscribe?: boolean;
+  shouldResubscribe?: any;
   children: (result: SubscriptionResult<TData>) => React.ReactNode;
 }
 
@@ -42,9 +42,10 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
   };
 
   static propTypes = {
-    query: PropTypes.object.isRequired,
+    subscription: PropTypes.object.isRequired,
     variables: PropTypes.object,
     children: PropTypes.func.isRequired,
+    shouldResubscribe: PropTypes.oneOfType([PropTypes.func, PropTypes.bool]),
   };
 
   private client: ApolloClient<any>;
@@ -71,10 +72,19 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
     nextProps: SubscriptionProps<TData, TVariables>,
     nextContext: SubscriptionContext,
   ) {
-    if (shallowEqual(this.props, nextProps) && this.client === nextContext.client) {
+    if (
+      shallowEqual(this.props.variables, nextProps.variables) &&
+      this.client === nextContext.client &&
+      this.props.subscription === nextProps.subscription
+    ) {
       return;
     }
-    const shouldNotResubscribe = this.props.shouldResubscribe === false;
+
+    let shouldResubscribe = nextProps.shouldResubscribe;
+    if (typeof shouldResubscribe === 'function') {
+      shouldResubscribe = !!shouldResubscribe(this.props, nextProps);
+    }
+    const shouldNotResubscribe = shouldResubscribe === false;
     if (this.client !== nextContext.client) {
       this.client = nextContext.client;
     }
@@ -105,7 +115,7 @@ class Subscription<TData = any, TVariables = any> extends React.Component<
   private initialize = (props: SubscriptionProps<TData, TVariables>) => {
     if (this.queryObservable) return;
     this.queryObservable = this.client.subscribe({
-      query: props.query,
+      query: props.subscription,
       variables: props.variables,
     });
   };