+++ test-examples.sh
@@ -1,0 +1,17 @@
+cd examples
+
+cd base
+
+npm install
+
+npm t
+
+cd ../
+
+cd components
+
+npm install 
+
+npm t
+
+cd ../../