+++ test/client/graphql/queries/errors.test.tsx
@@ -12,7 +12,6 @@ import {
   Query,
   QueryResult,
   DataValue,
-  query as queryHoc,
 } from '../../../../src';
 
 import stripSymbols from '../../../test-utils/stripSymbols';
@@ -239,152 +238,6 @@ describe('[queries] errors', () => {
     });
   });
 
-  it('will not log a warning when there is an error that is caught in the render method', () =>
-    new Promise((resolve, reject) => {
-      const query: DocumentNode = gql`
-        query people {
-          allPeople(first: 1) {
-            people {
-              name
-            }
-          }
-        }
-      `;
-
-      interface Data {
-        allPeople: {
-          people: { name: string }[];
-        };
-      }
-      const link = mockSingleLink({
-        request: { query },
-        error: new Error('oops'),
-      });
-      const client = new ApolloClient({
-        link,
-        cache: new Cache({ addTypename: false }),
-      });
-
-      const origError = console.error;
-      const errorMock = jest.fn();
-      console.error = errorMock;
-
-      let renderCount = 0;
-      @graphql<{}, Data>(query)
-      class HandledErrorComponent extends React.Component<ChildProps<{}, Data>> {
-        render(): React.ReactNode {
-          try {
-            switch (renderCount++) {
-              case 0:
-                expect(this.props.data!.loading).toEqual(true);
-                break;
-              case 1:
-                expect(this.props.data!.error!.message).toEqual('Network error: oops');
-                break;
-              default:
-                throw new Error('Too many renders.');
-            }
-          } catch (error) {
-            console.error = origError;
-            reject(error);
-          }
-          return null;
-        }
-      }
-
-      renderer.create(
-        <ApolloProvider client={client}>
-          <HandledErrorComponent />
-        </ApolloProvider>,
-      );
-
-      setTimeout(() => {
-        try {
-          expect(errorMock.mock.calls.length).toBe(0);
-          resolve();
-        } catch (error) {
-          reject(error);
-        } finally {
-          console.error = origError;
-        }
-      }, 20);
-    }));
-
-  it('will log a warning when there is an error that is not caught in the render method', () =>
-    new Promise((resolve, reject) => {
-      const query: DocumentNode = gql`
-        query people {
-          allPeople(first: 1) {
-            people {
-              name
-            }
-          }
-        }
-      `;
-
-      interface Data {
-        allPeople: {
-          people: { name: string }[];
-        };
-      }
-
-      const link = mockSingleLink({
-        request: { query },
-        error: new Error('oops'),
-      });
-      const client = new ApolloClient({
-        link,
-        cache: new Cache({ addTypename: false }),
-      });
-
-      const origError = console.error;
-      const errorMock = jest.fn();
-      console.error = errorMock;
-
-      let renderCount = 0;
-      @graphql<{}, Data>(query)
-      class UnhandledErrorComponent extends React.Component<ChildProps<{}, Data>> {
-        render(): React.ReactNode {
-          try {
-            switch (renderCount++) {
-              case 0:
-                expect(this.props.data!.loading).toEqual(true);
-                break;
-              case 1:
-                // Noop. Don’t handle the error so a warning will be logged to the console.
-                break;
-              default:
-                throw new Error('Too many renders.');
-            }
-          } catch (error) {
-            console.error = origError;
-            reject(error);
-          }
-          return null;
-        }
-      }
-
-      renderer.create(
-        <ApolloProvider client={client}>
-          <UnhandledErrorComponent />
-        </ApolloProvider>,
-      );
-
-      setTimeout(() => {
-        try {
-          expect(renderCount).toBe(2);
-          expect(errorMock.mock.calls.length).toBe(1);
-          expect(errorMock.mock.calls[0][0]).toEqual(
-            'Unhandled (in react-apollo:Apollo(UnhandledErrorComponent))',
-          );
-          resolve();
-        } catch (error) {
-          reject(error);
-        } finally {
-          console.error = origError;
-        }
-      }, 250);
-    }));
   it('will not log a warning when there is an error that is not caught in the render method when using query', () =>
     new Promise((resolve, reject) => {
       const query: DocumentNode = gql`
@@ -417,7 +270,7 @@ describe('[queries] errors', () => {
       console.error = errorMock;
 
       let renderCount = 0;
-      @queryHoc<{}, Data>(query)
+      @graphql<{}, Data>(query)
       class UnhandledErrorComponent extends React.Component<ChildProps<{}, Data>> {
         render(): React.ReactNode {
           try {
@@ -556,8 +409,8 @@ describe('[queries] errors', () => {
           try {
             switch (count++) {
               case 0:
-                props.data!
-                  .refetch()
+                props
+                  .data!.refetch()
                   .then(() => {
                     done.fail('Expected error value on first refetch.');
                   })
@@ -569,8 +422,8 @@ describe('[queries] errors', () => {
               case 2:
                 expect(props.data!.loading).toBeFalsy();
                 expect(props.data!.error).toBeTruthy();
-                props.data!
-                  .refetch()
+                props
+                  .data!.refetch()
                   .then(noop)
                   .catch(() => {
                     done.fail('Expected good data on second refetch.');
@@ -655,8 +508,8 @@ describe('[queries] errors', () => {
           try {
             switch (count++) {
               case 0:
-                props.data!
-                  .refetch()
+                props
+                  .data!.refetch()
                   .then(() => {
                     done.fail('Expected error value on first refetch.');
                   })
@@ -736,7 +589,6 @@ describe('[queries] errors', () => {
     });
 
     let count = 0;
-    const noop = () => null;
     type ContainerOwnProps = { toggle: () => void };
     const Container = graphql<ContainerOwnProps, Data>(query, {
       options: { notifyOnNetworkStatusChange: true },